backend/concert_generateur/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
packaging==25.0
pluggy==1.5.0
pycountry==24.6.1
pytest==8.3.5
pytest-django==4.11.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
pycountry==24.6.1
django_cors_headers==4.7.0
